moved dummy data to models, added more characters to GenId function
