+++ server/utils/common.js
@@ -1,4 +1,6 @@
-export function GenId(len=6, chars="abcdefghjkmnpqrstwxyz123456789"){
+export function GenId(
+  len=8, 
+  chars="abcdefghjkmnpqrtwxyABCDEFGHJKMNPQRTWXY123456789"){
   let id = "";
   while(len){
     id += chars[Math.random() * chars.length | 0];
